rename pong_repository -> ball_repository
