+++ app/domain/repository/ball_repository.ts
