+++ app/domain/repository/index.ts
@@ -1,5 +1,5 @@
+export * from "./ball_repository";
 export * from "./direct_message_repository";
 export * from "./friendship_repository";
-export * from "./pong_repository";
 export * from "./repository";
 export * from "./user_repository";
