+++ app/domain/repository/repository.ts
@@ -1,6 +1,6 @@
+import type { IBallRepository } from "./ball_repository";
 import type { IDirectMessageRepository } from "./direct_message_repository";
 import type { IFriendshipRepository } from "./friendship_repository";
-import type { IBallRepository } from "./pong_repository";
 import type { IUserRepository } from "./user_repository";
 
 export interface IRepository {
