+++ app/infra/kvs/index.ts
@@ -1,2 +1,2 @@
-export * from "./pong_repository";
+export * from "./ball_repository";
 export * from "./repository";
